section 8 (partial): split hfs_fsck Phase 1 (MDB) into submodule
Extract validate_hfs_name, check_alternate_mdb, check_embedded_hfs_plus,
check_mdb into src/fs/hfs_fsck/mdb.rs (260 lines). hfs_fsck/mod.rs drops
from 5278 to 5031 lines. All 61 fsck tests pass.
